Refactor server code
Changed the format for the commands that the server accepts.
Added validation for the new commands.
Also added a 'server root' which is where the files will be stored on the server.

These commands all correspond to the linux commands almost exactly.

For uploading files:

cp server://<filepath> <filepath>
This command is for downloading a file from the server to the client. This should fail if <filepath> already exists.
Only the client knows if <filepath> exists.
The client expects a message with the filesize first, then messages with file chunks.

cp -f server://<filepath> <filepath>
This command is the same as the above, but if <filepath> exists on the client, overwrite it with the file downloaded from the server
Only the server knows if server://<filepath> exists.

cp <filepath> server://<filepath>
This command is for uploading a file from the client to the server.
This should fail if server://<filepath> already exists.
The server expects a message with the filesize first, then messages with file chunks.

cp -f <filepath> server://<filepath>
This command is the same as the above, but if <filepath> exists on the server, overwrite it with the file uploaded from the client.

For deleting files and directories:

rm <filepath>
Deletes the file on the server at the path.

rm -d <filepath>
Deletes the directory on the server at the path. This fails if the directory is not empty.

Listing files:

ls <filepath>
The server will send a list of the files and directories at the filepath.

Making directories:

mkdir <filepath>
The server will try and create an empty directory at the filepath.